Search for a 404 template
When serving a static path, in case of 404 error instead of print
the error to the user, look for a 404.html template or index.html.
If none of them is founded the error is printed.

Also, remove tests for installation, to avoid waste on developers
devices. It is not necessari to be tested, since if npm install
fails on Travis, the entire test fails.