Implement #8 and #10: certificates in AppData at firts run.
This prevent the script to generate a new certificate every time
the module is updated.
`certs.js` manages the certificates: creation and deletion, and
provides them to index.js.
An uninstall script is added, it deletes the folder at uninstall
time.

makes problems, generate the certificates at first run.
If the index.js won't find the certificates,
certs.js will generate them.

Some edits in README accordingly to changes.
Add also a link on README to CONTRIBUTING.

Upgrade to v4.1.0 in package.json.
Upgrade of dependencies.